+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.6.9 - 2019-04-18
+
+### Added
+- Add getters for field attributes
+
+### Changed
+- Make private methods `protected`
+
 ## 0.6.8 - 2019-04-18
 
 ### Fixed